Cast gyms_using to int so it pluralizes correctly
Caught on the deployed equipment admin: rows read "1 gyms using". The
count came back from Neon as a string ("1"), and plural()'s strict n===1
check fails on a string, so it always rendered the plural. The old
equipment page happened to hide this by coercing with Number(); Kevin's
rebuild routed the count through plural() and exposed the latent missing
::int cast (the taxonomy API already casts, equipment never did).

Swept the other admin count(*) selects — this was the only uncast one.

+++ app/api/admin/equipment/route.ts
@@ -20,7 +20,7 @@ export async function GET() {
   const items = await sql`
     select c.id, c.name, c.normalized, c.status, c.created_at,
       t.name as proposed_by,
-      (select count(*) from tenant_equipment te where te.catalog_id = c.id) as gyms_using
+      (select count(*) from tenant_equipment te where te.catalog_id = c.id)::int as gyms_using
     from equipment_catalog c
     left join tenants t on t.id = c.created_by_tenant_id
     where c.status not in ('rejected', 'merged')
